[ITERATE] Animation support for turn in place / aim offset.
Turn in place rotation uses the difference (delta) between the current and previous actor rotation of the ABP's owning actor.
The rotation of the actor is interpolated to the cursor's location via TDSCharacter::LookAt() to face the cursor location. We only update the Yaw of the rotation, as we don't need to modify the Pitch or Roll, being a topdown-shooter.
The upper and lower body is now split within the ABP, which allows the use of the RotateRootBone node within the AnimGraph to counter rotate the skeletal mesh in opposite to the actor rotation. This is up until a threshold is reached within the AimOffset and a turn-in-place animation is played.
Within the upper and lower body split, it allows the blending of the locomotion and aiming poses, allowing the lower body to be counter rotated to remain in place, whilst the upper body can be rotated to aim towards the cursor location.