[Feature] Start of weapon system, focusing on hitscan weapons.
Upon equipping an item, a new equipment mapping context is added to the player, providing them with the UseItem keybinding (bound to left-click).
Prototype implementation of hitscan weapons implemented. Use the UseItem keybinding, if the player has an EquippedWeapon, we:
1. Play the character firing animation
2. Fire a line trace from the weapon's muzzle outwards to the weapon's range point
3. If the line trace hits an actor, we spawn an explosion effect at that location
Will need to be ported to C++ once out of prototype stage.
Added a new WeaponData property to specify the socket name the weapon should attach to when being equipped.