[FEATURE] Blended start/stop/cycle/pivot animations with hip fire override weight.
This commit allows the detailed animation states to blend with the hip fire pose when firing a weapon. Whilst moving, the lower body will continue to run the locomotion animations, whilst the upper body runs the hip fire pose animation, placing the weapon into the correct position for firing. 0.5 seconds after firing, the full body locomotion takes control again and the upper body will return to following the same animations as the lower body where required.
For this change, I had to know when the player was firing their weapon. To do this, I've iterated the prototype weapon system to set a bIsUsingItem flag when firing. At some point, this will need to be iterated to have the ABP instead react to some other kind of property being added/removed that's specifically for firing weapons.