[FEATURE] Implemented projectile and finished hitscan weapon prototypes.
Firing a weapon now runs the following logic:
1. If the player can fire the equipped weapon, play the character firing animation
2. Determine the UseMode of the weapon (projectile vs hitscan)
For hitscans, I've moved the previously implemented logic into a function. I've added an impulse if the hitscan hits something that simulates physics.
For projectiles, we spawn a new BP_Projectile actor slightly in front of the weapon's Muzzle socket. This projectile on impact spawns an impact VFX and fires an impulse using it's RadialForce component.
Created a new collision object response type for Projectiles as well. Removed the Config folder name from the gitignore file so this change was submitted.
Will need to be ported to C++ once out of prototype stage.