[ITERATE] Hitscan and Projectile weapon type iteration to end location and other general improvements.
Previously, for both hitscan and projectile weapons, the end location was determined firing out of the direction of the weapon muzzle. This could lead to inconsistent behaviour for where the bullet ended, which isn't ideal for a topdown shooter game. To fix this, the end location is now calculated by taking the Z axis of the weapon's muzzle socket location; creating a new vector using the Z axis value and the X and Y locations from the hit result under the cursor; adding that onto the forward vector of the character multiplied by the weapon's range. This means, that regardless of how the animation moves the weapon when firing, the bullet will always end up in the correct location for a topdown shooter.
Also turned off the tick by default for pickuppable items to improve performance.
Slightly rotated the Weapon socket within SK_Mannequin to improve the direction of the weapon in-game.
Added a custom event within ABP_Manny to toggle the debug information on/off using "ke * FuncName" console command.
Added new WeaponData properties for the character fire animation montage (so we can select different AnimMontages for different weapons) and the FireDelay time (which blocks the player from firing again during this time period).