[FEATURE] Implement burst weapon type.
New ranged weapon settings (only available when the FireMode is set to Burst) for BulletsPerBurst and BurstDelay. When shooting, the weapon will automatically fire the required BulletsPerBurst until the required bullets have been reached (or the weapon is out of ammo). Once this has been reached, a seperate delay is active (the BurstDelay) which delays the firing of the next burst of bullets. If the player stops holding the shoot button in this timeframe, the next burst will not be shot.

Testing:
PIE - Any map. Spawned a custom burst setup for testing that fired 3 bullets per burst, with an AttackDelay of 0.2f and a BurstDelay of 1.0f. Worked as intended.