[FIX] Weapon animations continue to run after dropping the equipped weapon.
Active weapon animations are now correctly ended when dropping the weapon.
Ported the play/end animation montage logic from TDSWeapon_Ranged to TDSWeaponBase. There is now a function for PlayMontage() which takes in a TFunction parameter that's used for the OnMontageEnded callback. This is primarily used for when the reload animation ends, as we require additional logic to run when this happens (actually performing the reload for the ammo count etc). For other animations like firing / dry firing etc we use a default callback that resets the ActiveAnimMontage if it hasn't been interrupted. If it's been interrupted, this is due to another montage taking over, so we don't want to set ActiveAnimMontage to nullptr.
I've also ported over the unlink of the weapon's anim class layer from blueprint to C++.
TDSWeaponBase also now holds a weak object pointer to the owner's anim instance, for easier use of the AnimInstance instead of having to retrieve it everytime the weapon wanted to play an animation.