[FEATURE] Create object management system.
Implemented a generic object pooling system within the CMGGameplayFramework plugin that can be used for any type of UObjects. If the object type passed through is an AActor, the object within the pool will be handled differently than base UObjects (for example, SpawnActor vs NewObject).
New settings page for the CMGGameplayFramework, which holds some settings for the object management, including cull thresholds, requested pools etc.
The projectiles within this project have been iterated to use the new object management system implemented within the plugin. 10 projectiles are created within the pool by default and we use the available projectiles when required by ranged projectile weapons. Projectiles are then released back to the pool instead of being destroyed once they hit something (this includes moving it back to 0,0,0; hiding it in-game; stopping it's movement; disabling it's collision; etc).